project-templates/quant/pyproject.toml
[project]
name = "my-quant-project"
version = "0.1.0"
description = ""
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "scikit-learn>=1.3.0",
    "plotly>=5.15.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "duckdb>=0.8.0",
    "pyarrow>=12.0.0",
    "PyYAML>=6.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "tqdm>=4.65.0",
    "openpyxl>=3.1.0",
    "statsmodels>=0.14.0",
    "tabulate>=0.9.0",
    "mlflow>=2.10.0",
    # Jupyter
    "jupyter>=1.0.0",
    "ipykernel>=6.0.0",
    "notebook>=7.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
